code/5/requirements.txt
baidu-aip
SpeechRecognition
pyaudio
